fixed lint issues

+++ spec/pg_sql_triggers/migrator_spec.rb
@@ -4,73 +4,71 @@ require "spec_helper"
 require "fileutils"
 
 RSpec.describe PgSqlTriggers::Migrator do
-  let(:migrations_path) { Rails.root.join("db", "triggers") }
+  let(:migrations_path) { Rails.root.join("db/triggers") }
 
   before do
-    FileUtils.mkdir_p(migrations_path) unless Dir.exist?(migrations_path)
+    FileUtils.mkdir_p(migrations_path)
   end
 
   after do
     # Clean up test migrations
-    if Dir.exist?(migrations_path)
-      Dir.glob(migrations_path.join("*.rb")).each { |f| File.delete(f) }
-    end
-    PgSqlTriggers::Migrator.ensure_migrations_table!
+    Dir.glob(migrations_path.join("*.rb")).each { |f| File.delete(f) } if Dir.exist?(migrations_path)
+    described_class.ensure_migrations_table!
     ActiveRecord::Base.connection.execute("TRUNCATE TABLE trigger_migrations")
   end
 
   describe ".migrations_path" do
     it "returns the correct path" do
-      expect(PgSqlTriggers::Migrator.migrations_path).to eq(Rails.root.join("db", "triggers"))
+      expect(described_class.migrations_path).to eq(Rails.root.join("db/triggers"))
     end
   end
 
   describe ".migrations_table_exists?" do
     it "returns true when table exists" do
-      PgSqlTriggers::Migrator.ensure_migrations_table!
-      expect(PgSqlTriggers::Migrator.migrations_table_exists?).to be true
+      described_class.ensure_migrations_table!
+      expect(described_class.migrations_table_exists?).to be true
     end
 
     it "returns false when table doesn't exist" do
       ActiveRecord::Base.connection.execute("DROP TABLE IF EXISTS trigger_migrations")
-      expect(PgSqlTriggers::Migrator.migrations_table_exists?).to be false
-      PgSqlTriggers::Migrator.ensure_migrations_table!
+      expect(described_class.migrations_table_exists?).to be false
+      described_class.ensure_migrations_table!
     end
   end
 
   describe ".ensure_migrations_table!" do
     it "creates migrations table if it doesn't exist" do
       ActiveRecord::Base.connection.execute("DROP TABLE IF EXISTS trigger_migrations")
-      expect(PgSqlTriggers::Migrator.migrations_table_exists?).to be false
+      expect(described_class.migrations_table_exists?).to be false
 
-      PgSqlTriggers::Migrator.ensure_migrations_table!
-      expect(PgSqlTriggers::Migrator.migrations_table_exists?).to be true
+      described_class.ensure_migrations_table!
+      expect(described_class.migrations_table_exists?).to be true
     end
 
     it "does nothing if table already exists" do
-      PgSqlTriggers::Migrator.ensure_migrations_table!
-      expect { PgSqlTriggers::Migrator.ensure_migrations_table! }.not_to raise_error
+      described_class.ensure_migrations_table!
+      expect { described_class.ensure_migrations_table! }.not_to raise_error
     end
   end
 
   describe ".current_version" do
     it "returns 0 when no migrations have been run" do
-      PgSqlTriggers::Migrator.ensure_migrations_table!
-      expect(PgSqlTriggers::Migrator.current_version).to eq(0)
+      described_class.ensure_migrations_table!
+      expect(described_class.current_version).to eq(0)
     end
 
     it "returns the latest migration version" do
-      PgSqlTriggers::Migrator.ensure_migrations_table!
+      described_class.ensure_migrations_table!
       ActiveRecord::Base.connection.execute("INSERT INTO trigger_migrations (version) VALUES ('20231215120001')")
       ActiveRecord::Base.connection.execute("INSERT INTO trigger_migrations (version) VALUES ('20231215120002')")
-      expect(PgSqlTriggers::Migrator.current_version).to eq(20231215120002)
+      expect(described_class.current_version).to eq(20_231_215_120_002)
     end
   end
 
   describe ".migrations" do
     it "returns empty array when migrations directory doesn't exist" do
-      FileUtils.rm_rf(migrations_path) if Dir.exist?(migrations_path)
-      expect(PgSqlTriggers::Migrator.migrations).to eq([])
+      FileUtils.rm_rf(migrations_path)
+      expect(described_class.migrations).to eq([])
     end
 
     it "parses migration files correctly" do
@@ -89,9 +87,9 @@ RSpec.describe PgSqlTriggers::Migrator do
       File.write(migrations_path.join("20231215120001_test_migration.rb"), migration_content)
       File.write(migrations_path.join("20231215120002_another_migration.rb"), migration_content)
 
-      migrations = PgSqlTriggers::Migrator.migrations
+      migrations = described_class.migrations
       expect(migrations.count).to eq(2)
-      expect(migrations.map(&:version)).to contain_exactly(20231215120001, 20231215120002)
+      expect(migrations.map(&:version)).to contain_exactly(20_231_215_120_001, 20_231_215_120_002)
       expect(migrations.map(&:name)).to contain_exactly("test_migration", "another_migration")
     end
 
@@ -107,8 +105,8 @@ RSpec.describe PgSqlTriggers::Migrator do
       File.write(migrations_path.join("20231215120001_first.rb"), migration_content)
       File.write(migrations_path.join("20231215120002_second.rb"), migration_content)
 
-      migrations = PgSqlTriggers::Migrator.migrations
-      expect(migrations.map(&:version)).to eq([20231215120001, 20231215120002, 20231215120003])
+      migrations = described_class.migrations
+      expect(migrations.map(&:version)).to eq([20_231_215_120_001, 20_231_215_120_002, 20_231_215_120_003])
     end
   end
 
@@ -127,26 +125,26 @@ RSpec.describe PgSqlTriggers::Migrator do
     end
 
     it "returns migrations with version greater than current" do
-      PgSqlTriggers::Migrator.ensure_migrations_table!
+      described_class.ensure_migrations_table!
       ActiveRecord::Base.connection.execute("INSERT INTO trigger_migrations (version) VALUES ('20231215120001')")
 
-      pending = PgSqlTriggers::Migrator.pending_migrations
-      expect(pending.map(&:version)).to eq([20231215120002, 20231215120003])
+      pending = described_class.pending_migrations
+      expect(pending.map(&:version)).to eq([20_231_215_120_002, 20_231_215_120_003])
     end
 
     it "returns all migrations when none have been run" do
-      PgSqlTriggers::Migrator.ensure_migrations_table!
-      pending = PgSqlTriggers::Migrator.pending_migrations
-      expect(pending.map(&:version)).to eq([20231215120001, 20231215120002, 20231215120003])
+      described_class.ensure_migrations_table!
+      pending = described_class.pending_migrations
+      expect(pending.map(&:version)).to eq([20_231_215_120_001, 20_231_215_120_002, 20_231_215_120_003])
     end
 
     it "returns empty array when all migrations are run" do
-      PgSqlTriggers::Migrator.ensure_migrations_table!
+      described_class.ensure_migrations_table!
       ActiveRecord::Base.connection.execute("INSERT INTO trigger_migrations (version) VALUES ('20231215120001')")
       ActiveRecord::Base.connection.execute("INSERT INTO trigger_migrations (version) VALUES ('20231215120002')")
       ActiveRecord::Base.connection.execute("INSERT INTO trigger_migrations (version) VALUES ('20231215120003')")
 
-      pending = PgSqlTriggers::Migrator.pending_migrations
+      pending = described_class.pending_migrations
       expect(pending).to be_empty
     end
   end
@@ -176,10 +174,10 @@ RSpec.describe PgSqlTriggers::Migrator do
       end
 
       it "applies all pending migrations" do
-        PgSqlTriggers::Migrator.ensure_migrations_table!
-        PgSqlTriggers::Migrator.run_up
+        described_class.ensure_migrations_table!
+        described_class.run_up
 
-        expect(PgSqlTriggers::Migrator.current_version).to eq(20231215120001)
+        expect(described_class.current_version).to eq(20_231_215_120_001)
         expect(ActiveRecord::Base.connection.table_exists?("test_table")).to be true
       end
     end
@@ -223,8 +221,8 @@ RSpec.describe PgSqlTriggers::Migrator do
       end
 
       it "applies only the specified migration" do
-        PgSqlTriggers::Migrator.ensure_migrations_table!
-        PgSqlTriggers::Migrator.run_up(20231215120002)
+        described_class.ensure_migrations_table!
+        described_class.run_up(20_231_215_120_002)
 
         version_exists = ActiveRecord::Base.connection.select_value(
           "SELECT 1 FROM trigger_migrations WHERE version = '20231215120002' LIMIT 1"
@@ -238,18 +236,18 @@ RSpec.describe PgSqlTriggers::Migrator do
       end
 
       it "raises error if migration doesn't exist" do
-        expect {
-          PgSqlTriggers::Migrator.run_up(99999999999999)
-        }.to raise_error(StandardError, /Migration version 99999999999999 not found/)
+        expect do
+          described_class.run_up(99_999_999_999_999)
+        end.to raise_error(StandardError, /Migration version 99999999999999 not found/)
       end
 
       it "raises error if migration already applied" do
-        PgSqlTriggers::Migrator.ensure_migrations_table!
+        described_class.ensure_migrations_table!
         ActiveRecord::Base.connection.execute("INSERT INTO trigger_migrations (version) VALUES ('20231215120001')")
 
-        expect {
-          PgSqlTriggers::Migrator.run_up(20231215120001)
-        }.to raise_error(StandardError, /already applied/)
+        expect do
+          described_class.run_up(20_231_215_120_001)
+        end.to raise_error(StandardError, /already applied/)
       end
     end
   end
@@ -271,8 +269,8 @@ RSpec.describe PgSqlTriggers::Migrator do
 
     before do
       File.write(migrations_path.join("20231215120001_test_migration.rb"), migration_content)
-      PgSqlTriggers::Migrator.ensure_migrations_table!
-      PgSqlTriggers::Migrator.run_up
+      described_class.ensure_migrations_table!
+      described_class.run_up
     end
 
     after do
@@ -281,14 +279,14 @@ RSpec.describe PgSqlTriggers::Migrator do
 
     it "rolls back the last migration" do
       expect(ActiveRecord::Base.connection.table_exists?("test_table")).to be true
-      PgSqlTriggers::Migrator.run_down
-      expect(PgSqlTriggers::Migrator.current_version).to eq(0)
+      described_class.run_down
+      expect(described_class.current_version).to eq(0)
     end
 
     it "returns early when no migrations exist" do
-      PgSqlTriggers::Migrator.run_down
-      expect(PgSqlTriggers::Migrator.current_version).to eq(0)
-      expect { PgSqlTriggers::Migrator.run_down }.not_to raise_error
+      described_class.run_down
+      expect(described_class.current_version).to eq(0)
+      expect { described_class.run_down }.not_to raise_error
     end
 
     context "when rolling back to specific version" do
@@ -323,19 +321,19 @@ RSpec.describe PgSqlTriggers::Migrator do
       before do
         File.write(migrations_path.join("20231215120002_second.rb"), second_migration_content)
         File.write(migrations_path.join("20231215120003_third.rb"), third_migration_content)
-        PgSqlTriggers::Migrator.run_up
+        described_class.run_up
       end
 
       it "rolls back to the specified version" do
-        expect(PgSqlTriggers::Migrator.current_version).to eq(20231215120003)
-        PgSqlTriggers::Migrator.run_down(20231215120002)
-        expect(PgSqlTriggers::Migrator.current_version).to eq(20231215120002)
+        expect(described_class.current_version).to eq(20_231_215_120_003)
+        described_class.run_down(20_231_215_120_002)
+        expect(described_class.current_version).to eq(20_231_215_120_002)
       end
 
       it "raises error if version not found or not applied" do
-        expect {
-          PgSqlTriggers::Migrator.run_down(99999999999999)
-        }.to raise_error(StandardError, /not found or not applied/)
+        expect do
+          described_class.run_down(99_999_999_999_999)
+        end.to raise_error(StandardError, /not found or not applied/)
       end
     end
   end
@@ -353,16 +351,16 @@ RSpec.describe PgSqlTriggers::Migrator do
     before do
       File.write(migrations_path.join("20231215120001_first.rb"), migration_content)
       File.write(migrations_path.join("20231215120002_second.rb"), migration_content)
-      PgSqlTriggers::Migrator.ensure_migrations_table!
+      described_class.ensure_migrations_table!
     end
 
     it "returns status for all migrations" do
       ActiveRecord::Base.connection.execute("INSERT INTO trigger_migrations (version) VALUES ('20231215120001')")
 
-      status = PgSqlTriggers::Migrator.status
+      status = described_class.status
       expect(status.count).to eq(2)
-      expect(status.find { |s| s[:version] == 20231215120001 }[:status]).to eq("up")
-      expect(status.find { |s| s[:version] == 20231215120002 }[:status]).to eq("down")
+      expect(status.find { |s| s[:version] == 20_231_215_120_001 }[:status]).to eq("up")
+      expect(status.find { |s| s[:version] == 20_231_215_120_002 }[:status]).to eq("down")
     end
   end
 
@@ -391,24 +389,23 @@ RSpec.describe PgSqlTriggers::Migrator do
       begin
         ActiveRecord::Base.connection.execute("CREATE FUNCTION existing_function() RETURNS TRIGGER AS $$ BEGIN RETURN NEW; END; $$ LANGUAGE plpgsql;")
         ActiveRecord::Base.connection.execute("CREATE TRIGGER existing_trigger BEFORE INSERT ON users FOR EACH ROW EXECUTE FUNCTION existing_function();")
-      rescue
+      rescue StandardError => _e
+        # Ignore errors - function/trigger may already exist
       end
     end
 
     after do
-      begin
-        ActiveRecord::Base.connection.execute("DROP TRIGGER IF EXISTS existing_trigger ON users")
-        ActiveRecord::Base.connection.execute("DROP FUNCTION IF EXISTS existing_function()")
-      rescue
-      end
+      ActiveRecord::Base.connection.execute("DROP TRIGGER IF EXISTS existing_trigger ON users")
+      ActiveRecord::Base.connection.execute("DROP FUNCTION IF EXISTS existing_function()")
+    rescue StandardError => _e
+      # Ignore errors during cleanup - trigger/function may not exist
     end
 
     it "removes registry entries for triggers that don't exist in database" do
       expect(PgSqlTriggers::TriggerRegistry.count).to eq(2)
-      PgSqlTriggers::Migrator.cleanup_orphaned_registry_entries
+      described_class.cleanup_orphaned_registry_entries
       expect(PgSqlTriggers::TriggerRegistry.count).to eq(1)
       expect(PgSqlTriggers::TriggerRegistry.first.trigger_name).to eq("existing_trigger")
     end
   end
 end
-
